+++ engine/Game.hpp
@@ -20,7 +20,7 @@ class Game {
 
         bool running() { return this->isRunning; }
 
-        static void AddTile(int id, int x, int y);
+        static void AddTile(int id, float x, float y);
         static SDL_Renderer *renderer;
         static SDL_Event event;
         static std::vector<ColliderComponent*> colliders;
